+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Bugfixes
 * (De)serializing of blank node Entities with Constrains did not keep the proper link
+* Allow to configure certain settings via `.iadopt-catalogue.yml`
 
 ## 0.1.0